[Bugfix][Store] Guard snapshot metadata format detection against replica_count overflow
Address review feedback on #2257. The format-detection arithmetic
(kFieldCount + replica_count) was done in uint32_t, so a near-UINT32_MAX
replica_count wraps around: e.g. with total_elements=7 and
replica_count=0xFFFFFFFF, 8 + 0xFFFFFFFF wraps to 7, making
is_one_extra_format true and letting an out-of-bounds index past the size
check.

Perform the comparison in uint64_t in both the standby restore reader
(catalog_backed_snapshot_provider.cpp) and the matching master-side parser
(master_service.cpp), which shares the identical pattern.

Add RejectsOverflowingReplicaCount test: a metadata entry packing zero
replicas but declaring UINT32_MAX is now rejected with DESERIALIZE_FAIL
instead of triggering an OOB read.

+++ mooncake-store/src/ha/snapshot/catalog_backed_snapshot_provider.cpp
@@ -136,10 +136,13 @@ DeserializeStandbyObjectMetadata(
         //   v1: 7 + replica_count, no data_type or trailing hard_pinned
         //   v2: 8 + replica_count, either data_type or trailing hard_pinned
         //   v3: 9 + replica_count, data_type plus trailing hard_pinned
-        constexpr uint32_t kOldFieldCount = 7;
-        constexpr uint32_t kOneExtraFieldCount = 8;
-        constexpr uint32_t kCurrentFieldCount = 9;
-        const uint32_t total_elements = object.via.array.size;
+        // Use 64-bit arithmetic: replica_count is attacker-controlled and a
+        // near-UINT32_MAX value would otherwise wrap the additions back into a
+        // valid-looking total and let an out-of-bounds index through.
+        constexpr uint64_t kOldFieldCount = 7;
+        constexpr uint64_t kOneExtraFieldCount = 8;
+        constexpr uint64_t kCurrentFieldCount = 9;
+        const uint64_t total_elements = object.via.array.size;
         const bool is_old_format =
             (total_elements == kOldFieldCount + replica_count);
         const bool is_one_extra_format =
